Claude-like-Architecture/claude-3.5-sonnet/requirements.txt
# Core dependencies
aiohttp>=3.8.0
anthropic==0.40.0
asyncio==3.4.3
autopep8==2.3.1
black==24.10.0
fastapi==0.115.6
langchain-anthropic==0.1.22
langchain-groq==0.1.6
langchain-openai==0.1.22
numpy>=1.21.0
openai==1.51.2
pandas>=1.3.0
psutil==6.1.0
pydantic==2.6.1
python-dateutil>=2.8.2
python-dotenv>=0.21.0
pyyaml>=6.0
requests==2.32.3
scikit-learn>=1.0.0
setuptools==57.4.0
slowapi==0.1.9
typing-extensions>=4.5.0
uvicorn==0.32.1

# Development dependencies
flake8>=4.0.0
isort>=5.10.0
mypy>=1.0.0
pre-commit>=2.20.0
pytest==8.3.4
pytest-cov>=4.0.0

# Documentation dependencies
myst-parser>=0.18.0
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0

# Performance dependencies
dask>=2023.0.0
ray>=2.0.0
